% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilite.R
\name{angle.plot}
\alias{angle.plot}
\title{Plot Fitted Results and Model Diagnostics}
\usage{
angle.plot(object, time)
}
\arguments{
\item{object}{a fitted joint mean correlation model of class "JmdlMod", i.e.,
typically the result of jmdl().}

\item{time}{a vector of obeservation time points}
}
\description{
plot (a) fitted angles; (b) fitted correlations vs time lag;
(c) the empirical distribution function vs the fitted distribution function;
(d) the empirical correlations vs the fitted correlations, when the descrete
longitudinal dataset is balanced.
}
