% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmdl.R
\docType{data}
\name{epilepsy2}
\alias{epilepsy2}
\title{epilepsy2 Data}
\format{A data frame with 280 rows and 7 variables}
\usage{
data(epilepsy2)
}
\description{
Data from a clinical trial of 59 epileptics. For a baseline, patients were observed for 8 weeks
and the number of seizures recorded. The patients were then randomized to treatment by the drug
Progabide (31 patients) or to the placebo group (28 patients). They were observed for four
2-week periods and the number of seizures recorded.
}
\details{
\itemize{
  \item id: identifying number
  \item treat: 1=treated, 0=not
  \item seizures: number of seizures
  \item expind: 0=baseline period, 1=treatment period
  \item timeadj: weeks of period
  \item age: in years
  \item time: in weeks
}
}
\keyword{datasets}
