% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{jmdlMod-class}
\alias{jmdlMod-class}
\title{Class "jmdlMod" of Fitted Joint Mean-Correlation Models.}
\description{
Class "jmdlMod" of Fitted Joint Mean-Correlation Models.
}
\section{Slots}{

\describe{
\item{\code{call}}{the matched call}

\item{\code{opt}}{the optimization result returned by optimizeJmdl}

\item{\code{args}}{arguments m, Y, X, W, time}

\item{\code{q}}{degree of polynomial of the time lag to model the lower triangular matrix}

\item{\code{std}}{standard error for parameters}

\item{\code{tval}}{t statistic components list}

\item{\code{p}}{p.value components list}

\item{\code{devcomp}}{the deviance components list}

\item{\code{offset}}{a vecter to be added to a linear predictor}
}}

