% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilite.R
\name{lrt.test}
\alias{lrt.test}
\title{Pairwise Likelihood Ratio Statistic Test}
\usage{
lrt.test(fit, id)
}
\arguments{
\item{fit}{a fitted joint mean correlation model of class "JmdlMod", i.e.,
typically the result of jmdl().}

\item{id}{the id of paremeter to test}
}
\description{
Conducts a pairwise likelihood ratio test for joint mean-correlation regression.
}
