% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmdl.R
\docType{data}
\name{pbc3}
\alias{pbc3}
\title{pbc3 Data}
\format{A data frame with 1390 rows and 15 variables}
\usage{
data(pbc3)
}
\description{
Followup of 312 randomised patients with primary biliary cirrhosis, a rare autoimmune
liver disease, at Mayo Clinic.
}
\details{
\itemize{
  \item id: patients identifier; in total there are 312 patients
  \item years: number of years between registration and the earlier of death, transplantion, or study analysis time
  \item status: a factor with levels alive, transplanted and dead
  \item drug: a factor with levels placebo and D-penicil
  \item age: at registration in years
  \item sex: a factor with levels male and female
  \item year: number of years between enrollment and this visit date, remaining values on the line of data refer to this visit
  \item ascites: a factor with levels No and Yes
  \item hepatomegaly: a factor with levels No and Yes
  \item spiders: a factor with levels No and Yes
  \item edema: a factor with levels No edema (i.e., no edema and no diuretic therapy for edema), edema no diuretics (i.e., edema present without diuretics, or edema resolved by diuretics), and edema despite diuretics (i.e., edema despite diuretic therapy)
  \item serBilir: serum bilirubin in mg/dl
  \item albumin: albumin in gm/dl
  \item platelets: platelets per cubic ml / 1000
  \item prothrombin: prothrombin time in seconds
}
}
\keyword{datasets}
