% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmdl.R
\docType{methods}
\name{show,jmdlMod-method}
\alias{show,jmdlMod-method}
\title{Print information for jmdlMod-class}
\usage{
\S4method{show}{jmdlMod}(object)
}
\arguments{
\item{object}{a fitted joint mean correlation model of class "jmdlMod", i.e.,
typically the result of jmdl().}
}
\description{
Print information for jmdlMod-class
}
