% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmdl.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{toy Data}
\format{A data frame with 50 rows and 4 variables}
\usage{
data(toydata)
}
\description{
A simulation dataset.
}
\details{
\itemize{
  \item id: identifying number
  \item Y: response variable: 1, 0
  \item time: time
  \item X: independent variable
}
}
\keyword{datasets}
