% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilite.R
\name{un.angle.plot}
\alias{un.angle.plot}
\title{Plot Fitted Results}
\usage{
un.angle.plot(object)
}
\arguments{
\item{object}{a fitted joint mean correlation model of class "JmdlMod", i.e.,
typically the result of jmdl().}
}
\description{
plot (a) fitted angles; (b) fitted correlations vs time lag, when
the descrete longitudinal dataset is unbalanced.
}
